Add DuplicateCourseRedirect concern for CourseChoices
