+++ app/controllers/candidate_interface/continuous_applications/course_choices/course_study_mode_controller.rb
@@ -2,7 +2,7 @@ module CandidateInterface
   module ContinuousApplications
     module CourseChoices
       class CourseStudyModeController < BaseController
-        before_action :redirect_duplicate, only: %w[new] # rubocop:disable Rails/LexicallyScopedActionFilter
+        include Concerns::DuplicateCourseRedirect
 
       private
 
@@ -22,12 +22,6 @@ module CandidateInterface
 
         def current_step
           :course_study_mode
-        end
-
-        def redirect_duplicate
-          return unless current_application.contains_course?(Course.find(params[:course_id]))
-
-          redirect_to candidate_interface_continuous_applications_duplicate_course_selection_path(params[:provider_id], params[:course_id])
         end
       end
     end
