+++ app/controllers/candidate_interface/continuous_applications/course_choices/find_course_selection_controller.rb
@@ -2,19 +2,12 @@ module CandidateInterface
   module ContinuousApplications
     module CourseChoices
       class FindCourseSelectionController < BaseController
-        before_action :redirect_duplicate, only: %w[new] # rubocop:disable Rails/LexicallyScopedActionFilter
+        include Concerns::DuplicateCourseRedirect
 
       private
 
         def current_step
           :find_course_selection
-        end
-
-        def redirect_duplicate
-          course = Course.find(params[:course_id])
-          return unless current_application.contains_course?(course)
-
-          redirect_to candidate_interface_continuous_applications_duplicate_course_selection_path(course.provider_id, course.id)
         end
       end
     end
